+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/di/KoinModule.kt
@@ -2,7 +2,9 @@ package com.altmann.choresmanager.di
 
 import com.altmann.choresmanager.database.UserDao
 import com.altmann.choresmanager.local.LocalDatabase
-import com.altmann.choresmanager.ui.screens.home.HomeViewModel
+import com.altmann.choresmanager.ui.screens.calendar.CalendarViewModel
+import com.altmann.choresmanager.ui.screens.sidebar.SideBarViewModel
+import com.altmann.choresmanager.viewmodels.SharedChoreViewModel
 import org.koin.core.KoinApplication
 import org.koin.core.context.startKoin
 import org.koin.core.module.Module
@@ -17,7 +19,11 @@ val sharedModule = module {
         UserDao(get())
     }
     single<LocalDatabase> { LocalDatabase(get()) }
-    viewModelOf(::HomeViewModel)
+    single<SharedChoreViewModel> {
+        SharedChoreViewModel()
+    }
+    viewModelOf(::CalendarViewModel)
+    viewModelOf(::SideBarViewModel)
 
 }