+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/college/CollegeChore.kt
@@ -10,19 +10,7 @@ import kotlinx.datetime.LocalTime
 import kotlin.math.floor
 
 class CollegeChore(
-    choreId: Int,
-    startTime: LocalTime,
-    endTime: LocalTime,
-    daysOfWeek: List<DayOfWeek>,
-    startDate: LocalDate,
-    endDate: LocalDate,
-    choreException: List<LocalDate>,
-    title: String,
-    description: String,
-    priority: Priority,
-    deleted: Boolean? = false,
-    finished: Boolean? = false,
-    color: Color = Color.Blue,
+    chore: Chore,
 
     val subject: String,
     val location: String?,
@@ -32,9 +20,9 @@ class CollegeChore(
     var absencesAllowed: Int = 0,
     var absencesCurrent: Int = 0
 ) : Chore(
-    choreId, startTime, endTime, daysOfWeek,
-    startDate, endDate, choreException, title, description,
-    priority, deleted, finished, color
+    chore.choreId, chore.startTime, chore.endTime, chore.daysOfWeek,
+    chore.startDate, chore.endDate, chore.choreException, chore.title, chore.description,
+    chore.priority, chore.deleted, chore.finished, chore.color
 ) {
 
     init {